# StayLocal_OpenAPI.yaml
openapi: 3.0.3
info:
  title: StayLocal Backend API
  version: 1.0.0
  description: Contract-first REST API for the StayLocal community tourism platform.
    Covers authentication, destinations, properties, media, camera+GPS verification,
    bookings, rewards, reviews and admin moderation.
  contact:
    name: StayLocal Project Team
    email: staylocal-team@example.com
servers:
- url: http://localhost:8787
  description: Local development
- url: http://localhost:8787
  description: Demo/CI default
...
paths:
...
  /api/v1/destinations:
    get:
      operationId: listDestinations
      summary: List StayLocal destinations
      description: Returns the four seeded destinations used by the project demo.
      tags:
      - Destinations
      security: *id001
      responses:
        '200':
          description: Destination list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Destination'
              examples:
                200_destinations_success:
                  value:
                  - &id004
                    id: dest-bhimashankar
                    name: Bhimashankar
                    slug: bhimashankar
                    latitude: 19.0729
                    longitude: 73.5358
                    description: Forest, temple and nature-focused destination.
                    imageUrl: https://example.com/bhimashankar.jpg
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                500_destinations_server_error:
                  value:
                    code: INTERNAL_ERROR
                    message: Unexpected server error.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/destinations/{destinationId}:
    get:
      operationId: getDestination
      summary: Get a destination
      description: Returns a single destination by ID.
      tags:
      - Destinations
      security: *id001
      parameters:
      - name: destinationId
        in: path
        required: true
        schema:
          type: string
          minLength: 3
          maxLength: 80
        examples:
          200_destination_bhimashankar:
            value: dest-bhimashankar
          404_destination_missing:
            value: dest-missing
      responses:
        '200':
          description: Destination found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Destination'
              examples:
                200_destination_bhimashankar:
                  value: *id004
        '404':
          description: Destination not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_destination_missing:
                  value:
                    code: NOT_FOUND
                    message: Destination not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Destination ID is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_destination_invalid_id:
                  value:
                    code: VALIDATION_ERROR
                    message: Destination ID is invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/properties:
    get:
      operationId: listProperties
      summary: Search approved properties
      description: Returns public, approved listings with price and destination filters
        and optional date availability filtering. Results are paginated.
      tags:
      - Properties
      security: *id001
      parameters:
      - name: destinationId
        in: query
        required: false
        schema:
          type: string
        examples:
          200_properties_filtered:
            value: dest-bhimashankar
      - name: minPrice
        in: query
        required: false
        schema:
          type: number
          minimum: 0
        examples:
          200_properties_filtered:
            value: 1000
      - name: maxPrice
        in: query
        required: false
        schema:
          type: number
          minimum: 0
        examples:
          200_properties_filtered:
            value: 3000
      - name: checkIn
        in: query
        required: false
        schema:
          type: string
          format: date
        examples:
          200_properties_available:
            value: '2026-10-10'
      - name: checkOut
        in: query
        required: false
        schema:
          type: string
          format: date
        examples:
          200_properties_available:
            value: '2026-10-12'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
        examples:
          200_properties_paged:
            value: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
        examples:
          200_properties_paged:
            value: 12
      responses:
        '200':
          description: Property results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedProperties'
              examples:
                200_properties_filtered:
                  value:
                    items:
                    - &id005
                      id: prop-1001
                      hostId: host-1001
                      destinationId: dest-bhimashankar
                      title: Forest View Local Homestay
                      description: Simple private room with a local host near Bhimashankar.
                      pricePerNight: 1800
                      maxGuests: 3
                      amenities:
                      - Wi-Fi
                      - Parking
                      - Hot Water
                      photos: []
                      status: APPROVED
                      latitude: 19.0731
                      longitude: 73.5356
                      locationVerified: true
                      verification: null
                      createdAt: '2026-09-20T09:00:00Z'
                      updatedAt: '2026-09-20T09:00:00Z'
                    page: 1
                    pageSize: 12
                    total: 1
        '422':
          description: One or more search parameters are invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_properties_invalid_filter:
                  value:
                    code: VALIDATION_ERROR
                    message: One or more search parameters are invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '429':
          description: Request rate exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                429_properties_rate_limited:
                  value:
                    code: RATE_LIMITED
                    message: Request rate exceeded.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                500_properties_server_error:
                  value:
                    code: INTERNAL_ERROR
                    message: Unexpected server error.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
    post:
      operationId: createProperty
      summary: Create a room listing
      description: Creates a draft listing for an approved host. A non-approved host
        cannot publish or submit a listing.
      tags:
      - Properties
      security: *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - destinationId
              - title
              - description
              - pricePerNight
              - maxGuests
              - latitude
              - longitude
              properties:
                destinationId:
                  type: string
                title:
                  type: string
                  minLength: 3
                  maxLength: 160
                description:
                  type: string
                  minLength: 20
                  maxLength: 5000
                pricePerNight:
                  type: number
                  minimum: 1
                  maximum: 100000
                maxGuests:
                  type: integer
                  minimum: 1
                  maximum: 20
                amenities:
                  type: array
                  items:
                    type: string
                  maxItems: 30
                latitude:
                  type: number
                  minimum: -90
                  maximum: 90
                longitude:
                  type: number
                  minimum: -180
                  maximum: 180
            examples:
              201_property_create_success:
                value:
                  destinationId: dest-bhimashankar
                  title: Forest View Local Homestay
                  description: Simple private room with a local host near Bhimashankar.
                  pricePerNight: 1800
                  maxGuests: 3
                  amenities:
                  - Wi-Fi
                  - Parking
                  - Hot Water
                  latitude: 19.0731
                  longitude: 73.5356
              422_property_create_invalid_price:
                value:
                  destinationId: dest-bhimashankar
                  title: Bad Property
                  description: This description is long enough for validation.
                  pricePerNight: 0
                  maxGuests: 3
                  latitude: 19.0731
                  longitude: 73.5356
      responses:
        '201':
          description: Property created as DRAFT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
              examples:
                201_property_create_success:
                  value: *id005
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_property_create_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: Only an approved host can create listings.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_property_create_forbidden:
                  value:
                    code: FORBIDDEN
                    message: Only an approved host can create listings.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_property_create_invalid_price:
                  value:
                    code: VALIDATION_ERROR
                    message: Request validation failed.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                500_property_create_server_error:
                  value:
                    code: INTERNAL_ERROR
                    message: Unexpected server error.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/properties/{propertyId}:
    get:
      operationId: getProperty
      summary: Get property details
      description: Returns a property, host-facing verification information, and public
        photos. Rejected/inactive properties are hidden from public users.
      tags:
      - Properties
      security: *id001
      parameters:
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
        examples:
          200_property_details:
            value: prop-1001
          404_property_missing:
            value: prop-missing
      responses:
        '200':
          description: Property found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
              examples:
                200_property_details:
                  value: *id005
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_property_missing:
                  value:
                    code: NOT_FOUND
                    message: Property not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
    patch:
      operationId: updateProperty
      summary: Edit a room listing
      description: Updates a host-owned listing. Changes to material listing data
        reset approval to PENDING so the admin can review the new state.
      tags:
      - Properties
      security: *id003
      parameters:
      - name: propertyId
        in: path
        required: true
        schema:
          type: string
        examples:
          200_property_update_success:
            value: prop-1001
          404_property_update_missing:
            value: prop-missing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 160
                description:
                  type: string
                  minLength: 20
                  maxLength: 5000
                pricePerNight:
                  type: number
                  minimum: 1
                  maximum: 100000
                maxGuests:
                  type: integer
                  minimum: 1
                  maximum: 20
                amenities:
                  type: array
                  items:
                    type: string
            examples:
              200_property_update_success:
                value:
                  pricePerNight: 1900
                  maxGuests: 3
              422_property_update_invalid:
                value:
                  pricePerNight: 0
      responses:
        '200':
          description: Property updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Property'
              examples:
                200_property_update_success:
                  value: *id005
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_property_update_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: User does not own the property.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_property_update_forbidden:
                  value:
                    code: FORBIDDEN
                    message: User does not own the property.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '404':
          description: Property not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_property_update_missing:
                  value:
                    code: NOT_FOUND
                    message: Property not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_property_update_invalid:
                  value:
                    code: VALIDATION_ERROR
                    message: Request validation failed.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
...
  /api/v1/bookings:
    get:
      operationId: listMyBookings
      summary: List current user's bookings
      description: Tourists see their own bookings. Hosts see bookings for their approved
        listings. Admins see all bookings.
      tags:
      - Bookings
      security: *id003
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - PENDING
          - CONFIRMED
          - CANCELLED
          - COMPLETED
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          maximum: 50
          default: 12
      responses:
        '200':
          description: Booking list
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - page
                - pageSize
                - total
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Booking'
                  page:
                    type: integer
                  pageSize:
                    type: integer
                  total:
                    type: integer
              examples:
                200_bookings_success:
                  value:
                    items:
                    - &id006
                      id: book-1001
                      propertyId: prop-1001
                      touristId: 10000000-0000-4000-8000-000000000001
                      checkIn: '2026-10-10'
                      checkOut: '2026-10-12'
                      guests: 2
                      nightCount: 2
                      subtotal: 3600
                      discount: 100
                      total: 3500
                      couponCode: STAY100-AB12CD
                      status: CONFIRMED
                      paymentMethod: UPI_QR
                      createdAt: '2026-09-28T12:00:00Z'
                      confirmedAt: '2026-09-28T12:01:00Z'
                    page: 1
                    pageSize: 12
                    total: 1
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_bookings_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
    post:
      operationId: createBooking
      summary: Create a booking
      description: Creates a booking after server-side date, guest, price and overlap
        validation. A booking can include an available ₹100 reward coupon from a prior
        booking.
      tags:
      - Bookings
      security: *id003
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - propertyId
              - checkIn
              - checkOut
              - guests
              - paymentMethod
              properties:
                propertyId:
                  type: string
                checkIn:
                  type: string
                  format: date
                checkOut:
                  type: string
                  format: date
                guests:
                  type: integer
                  minimum: 1
                  maximum: 20
                couponCode:
                  type: string
                  maxLength: 40
                  nullable: true
                paymentMethod:
                  type: string
                  enum:
                  - UPI_QR
                  - DEMO_CARD
                  - PAY_AT_STAY
            examples:
              201_booking_success:
                value:
                  propertyId: prop-1001
                  checkIn: '2026-10-10'
                  checkOut: '2026-10-12'
                  guests: 2
                  couponCode: STAY100-AB12CD
                  paymentMethod: UPI_QR
              409_booking_overlap:
                value:
                  propertyId: prop-1001
                  checkIn: '2026-10-11'
                  checkOut: '2026-10-13'
                  guests: 2
                  paymentMethod: PAY_AT_STAY
      responses:
        '201':
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
              examples:
                201_booking_success:
                  value: *id006
        '400':
          description: Malformed booking request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                400_booking_bad_request:
                  value:
                    code: BAD_REQUEST
                    message: Malformed booking request.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_booking_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '404':
          description: Property or coupon not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_booking_dependency_missing:
                  value:
                    code: NOT_FOUND
                    message: Property or coupon not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '409':
          description: Selected dates overlap an existing booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                409_booking_overlap:
                  value:
                    code: BOOKING_CONFLICT
                    message: Selected dates overlap an existing booking.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '422':
          description: Check-out must be after check-in, guest count is too high,
            or coupon is invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                422_booking_validation:
                  value:
                    code: BOOKING_INVALID
                    message: Check-out must be after check-in, guest count is too
                      high, or coupon is invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
  /api/v1/bookings/{bookingId}:
    get:
      operationId: getBooking
      summary: Get booking details
      description: Returns a booking visible to its tourist, listing host, or an admin.
      tags:
      - Bookings
      security: *id003
      parameters:
      - name: bookingId
        in: path
        required: true
        schema:
          type: string
        examples:
          200_booking_details:
            value: book-1001
          404_booking_missing:
            value: book-missing
      responses:
        '200':
          description: Booking found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
              examples:
                200_booking_details:
                  value: *id006
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_booking_details_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '403':
          description: User cannot access this booking.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                403_booking_details_forbidden:
                  value:
                    code: FORBIDDEN
                    message: User cannot access this booking.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
        '404':
          description: Booking not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                404_booking_missing:
                  value:
                    code: NOT_FOUND
                    message: Booking not found.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
...
  /api/v1/rewards:
    get:
      operationId: listRewards
      summary: List current user's reward coupons
      description: Returns available, redeemed and expired ₹100 reward coupons for
        the authenticated tourist.
      tags:
      - Rewards
      security: *id003
      responses:
        '200':
          description: Reward list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RewardCoupon'
              examples:
                200_rewards_success:
                  value:
                  - &id010
                    id: rew-1001
                    code: STAY100-AB12CD
                    amount: 100
                    currency: INR
                    status: AVAILABLE
                    issuedForBookingId: book-1001
                    redeemedOnBookingId: null
                    expiresAt: null
                    createdAt: '2026-09-28T12:01:00Z'
        '401':
          description: Authentication token is missing or invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardError'
              examples:
                401_rewards_unauthorized:
                  value:
                    code: UNAUTHORIZED
                    message: Authentication token is missing or invalid.
                    requestId: 11111111-1111-4111-8111-111111111111
                    details: {}
...
components:
  schemas:
    StandardError:
      type: object
      required:
      - code
      - message
      - requestId
      properties:
        code:
          type: string
          example: BOOKING_CONFLICT
        message:
          type: string
          example: The selected dates are already booked.
        requestId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        details:
          type: object
          additionalProperties: true
          nullable: true
      example:
        code: BOOKING_CONFLICT
        message: The selected dates are already booked.
        requestId: 11111111-1111-4111-8111-111111111111
        details:
          propertyId: prop-1001
...
    Destination:
      type: object
      required:
      - id
      - name
      - slug
      - latitude
      - longitude
      - description
      properties:
        id:
          type: string
          example: dest-bhimashankar
        name:
          type: string
          maxLength: 120
        slug:
          type: string
          pattern: ^[a-z0-9-]+$
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        description:
          type: string
          maxLength: 2000
        imageUrl:
          type: string
          format: uri
          nullable: true
      example: *id004
...
    PaginatedProperties:
      type: object
      required:
      - items
      - page
      - pageSize
      - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Property'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT